Support updating values averaged over a specific timespan
Currently the average of the last 5s for all sensor values is being
returned. This can result to missleading results when e.g. updating only
every 30s: a peak of 1000W might get missed and not accounted for.

Support custom timespans and use a default timespan of 30s to match
what the integrations is using.